Create BMI-Calc
The BMI calculator shows some simple mathematics that assign onto variables. I used the nested if-else structure for this app to demonstrate it.

It asks for the users height, age, then calculates the BMI using a formula to then display whether the entered info is above, within, or under the BMI.

Perfect for beginners/juniors!